refactor(mjh/discover): one component per file (quality gate)
Extract SavedSearchRow and DiscoverTabButton to their own files so each
module exports exactly one component. Resolves the inline-component
quality gate violation.
