+++ apps/myjobhunter/frontend/src/features/discover/DiscoverViewTabs.tsx
@@ -1,11 +1,15 @@
+import DiscoverTabButton from "@/features/discover/DiscoverTabButton";
 import type { DiscoverView } from "@/types/discovery/discover-view";
 
-interface ViewTabsProps {
+interface DiscoverViewTabsProps {
   activeView: DiscoverView;
   onSelect: (view: DiscoverView) => void;
 }
 
-export default function DiscoverViewTabs({ activeView, onSelect }: ViewTabsProps) {
+export default function DiscoverViewTabs({
+  activeView,
+  onSelect,
+}: DiscoverViewTabsProps) {
   return (
     <div
       role="tablist"
@@ -25,31 +29,5 @@ export default function DiscoverViewTabs({ activeView, onSelect }: ViewTabsProps
         onSelect={onSelect}
       />
     </div>
-  );
-}
-
-interface TabButtonProps {
-  label: string;
-  view: DiscoverView;
-  activeView: DiscoverView;
-  onSelect: (view: DiscoverView) => void;
-}
-
-function DiscoverTabButton({ label, view, activeView, onSelect }: TabButtonProps) {
-  const isActive = view === activeView;
-  return (
-    <button
-      role="tab"
-      aria-selected={isActive}
-      onClick={() => onSelect(view)}
-      className={[
-        "px-4 py-2 text-sm font-medium border-b-2 -mb-px transition-colors",
-        isActive
-          ? "border-primary text-foreground"
-          : "border-transparent text-muted-foreground hover:text-foreground hover:border-border",
-      ].join(" ")}
-    >
-      {label}
-    </button>
   );
 }
